WIP: Putting together core ioctl module (probably move to nix once stable)
None of this code is tested at this point, but the basics seem to be
starting to come together.  Ioctls are a mess to work with since there
is very little restriction on how they can be used.  There are however
some common patterns which is what this implementation will focus on
making much easier to use.

It is amazing how much is required to get to the point where writing
a function like this is possible:

    pub fn spi_ioc_rd_mode(fd: RawFd) -> io::Result<u8> {
        spidev_ioc_read(fd, SPI_IOC_NR_MODE)
    }