spidev: fix bug preventing bytes from being transmitted
Vec::clone_from_slice will copy the minimum of the length of the
vector and the length of the slice into the vector.  Since the lenght
of the vector was 0, no bytes were being copied from the tx_buf slice
into the tx_buf_vec.

The loopback test now is working as expected.

Signed-off-by: Paul Osborne <osbpau@gmail.com>